Added API documentation and fixed some bugs.
