+++ AGXUnity/IO/URDF/Geometry.cs
@@ -4,8 +4,28 @@ using UnityEngine;
 
 namespace AGXUnity.IO.URDF
 {
+  /// <summary>
+  /// Required element "geometry" for optional elements "collision" and "visual".
+  /// This element reads:
+  ///   - Either required element; "box", "cylinder", "sphere" or "mesh".
+  ///     - "box":      Required attribute "size".
+  ///     - "cylinder": Required attributes "radius" and "length".
+  ///     - "sphere":   Required attribute "radius".
+  ///     - "mesh":     Required attribute "filename" and optional attribute "scale" (default [1, 1, 1]).
+  /// </summary>
+  /// <remarks>
+  /// This class will throw an UrdfIOException when used properties doesn't match
+  /// the geometry type read. E.g., will throw when Type == GeometryType.Box
+  /// and Geometry.Radius is used.
+  /// </remarks>
   public class Geometry : Element
   {
+    /// <summary>
+    /// Reads data where "geometry" is required. Throws an UrdfIOException
+    /// if "geometry" isn't an element to given <paramref name="parent"/>.
+    /// </summary>
+    /// <param name="parent">Parent element "visual" or "collision".</param>
+    /// <returns>Read geometry instance.</returns>
     public static Geometry ReadRequired( XElement parent )
     {
       var geometryElement = parent.Element( "geometry" );
@@ -14,8 +34,34 @@ namespace AGXUnity.IO.URDF
       return new Geometry( geometryElement );
     }
 
-    public enum GeometryType { Box, Cylinder, Sphere, Mesh, Unknown }
+    /// <summary>
+    /// Geometry types given specification.
+    /// </summary>
+    public enum GeometryType
+    {
+      /// <summary>
+      /// Geometry type "box" with required attribute "size" => FullExtents.
+      /// </summary>
+      Box,
+      /// <summary>
+      /// Geometry type "cylinder" with required attributes "radius" => Radius and "length" => Length.
+      /// </summary>
+      Cylinder,
+      /// <summary>
+      /// Geometry type "sphere" with required attribute "radius" => Radius.
+      /// </summary>
+      Sphere,
+      /// <summary>
+      /// Geometry type "mesh" with required attribute "filename" => Filename and
+      /// optional attribute "scale" => Scale (default [1, 1, 1]).
+      /// </summary>
+      Mesh,
+      Unknown
+    }
 
+    /// <summary>
+    /// GeometryType.Mesh scale.
+    /// </summary>
     public Vector3 Scale
     {
       get
@@ -32,6 +78,9 @@ namespace AGXUnity.IO.URDF
       }
     }
 
+    /// <summary>
+    /// GeometryType.Mesh filename.
+    /// </summary>
     public string Filename
     {
       get
@@ -48,6 +97,9 @@ namespace AGXUnity.IO.URDF
       }
     }
 
+    /// <summary>
+    /// GeometryType.Box full extents (size).
+    /// </summary>
     public Vector3 FullExtents
     {
       get
@@ -64,6 +116,9 @@ namespace AGXUnity.IO.URDF
       }
     }
 
+    /// <summary>
+    /// GeometryType.Cylinder and GeometryType.Sphere radius.
+    /// </summary>
     public float Radius
     {
       get
@@ -80,6 +135,9 @@ namespace AGXUnity.IO.URDF
       }
     }
 
+    /// <summary>
+    /// GeometryType.Cylinder length.
+    /// </summary>
     public float Length
     {
       get
@@ -96,11 +154,18 @@ namespace AGXUnity.IO.URDF
       }
     }
 
+    /// <summary>
+    /// Type of this geometry.
+    /// </summary>
     public GeometryType Type { get; private set; } = GeometryType.Unknown;
 
+    /// <summary>
+    /// Reads required element "geometry".
+    /// </summary>
+    /// <param name="element">Required "geometry" element - invalid if null.</param>
+    /// <param name="optional">Unused.</param>
     public override void Read( XElement element, bool optional = true )
     {
-      base.Read( element, true );
       var children = element.Elements().ToArray();
       if ( children.Length != 1 )
         throw new UrdfIOException( $"{Utils.GetLineInfo( element )}: Invalid 'geometry' - expecting 1 geometry type, got {children.Length}." );
@@ -131,6 +196,10 @@ namespace AGXUnity.IO.URDF
       }
     }
 
+    /// <summary>
+    /// Construct given "geometry" element.
+    /// </summary>
+    /// <param name="element">Required "geometry" element - invalid if null.</param>
     public Geometry( XElement element )
     {
       Read( element );
