+++ AGXUnity/IO/URDF/Inertia.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e54970d778460414082a41e531374bec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
