+++ AGXUnity/IO/URDF/Inertial.cs
@@ -1,85 +1,23 @@
 ﻿using System.Xml.Linq;
-using UnityEngine;
 
 namespace AGXUnity.IO.URDF
 {
-  public struct Inertia
-  {
-    public static readonly Inertia zero;
-
-    public static Inertia Read( XElement element )
-    {
-      if ( element == null )
-        return zero;
-
-      float xx = (float)element.Attribute( "ixx" );
-      float xy = (float)element.Attribute( "ixy" );
-      float xz = (float)element.Attribute( "ixz" );
-      float yy = (float)element.Attribute( "iyy" );
-      float yz = (float)element.Attribute( "iyz" );
-      float zz = (float)element.Attribute( "izz" );
-
-      var inertia = new Inertia();
-      inertia[ 0, 0 ] = xx;
-      inertia[ 1, 1 ] = yy;
-      inertia[ 2, 2 ] = zz;
-      inertia[ 0, 1 ] = xy;
-      inertia[ 0, 2 ] = xz;
-      inertia[ 1, 2 ] = yz;
-
-      return inertia;
-    }
-
-    public static implicit operator agx.SPDMatrix3x3( Inertia inertia )
-    {
-      return new agx.SPDMatrix3x3( inertia[ 0, 0 ], inertia[ 0, 1 ], inertia[ 0, 2 ],
-                                   inertia[ 1, 0 ], inertia[ 1, 1 ], inertia[ 1, 2 ],
-                                   inertia[ 2, 0 ], inertia[ 2, 1 ], inertia[ 2, 2 ] );
-    }
-
-    public static explicit operator agx.Matrix3x3( Inertia inertia )
-    {
-      return new agx.Matrix3x3( inertia[ 0, 0 ], inertia[ 0, 1 ], inertia[ 0, 2 ],
-                                inertia[ 1, 0 ], inertia[ 1, 1 ], inertia[ 1, 2 ],
-                                inertia[ 2, 0 ], inertia[ 2, 1 ], inertia[ 2, 2 ] );
-    }
-
-    public float this[ int i, int j ]
-    {
-      get
-      {
-        var index = GetIndex( i, j );
-        return index.y == 0 ? m_data1[ index.x ] : m_data2[ index.x ];
-      }
-      set
-      {
-        var index = GetIndex( i, j );
-        if ( index.y == 0 )
-          m_data1[ index.x ] = value;
-        else
-          m_data2[ index.x ] = value;
-      }
-    }
-
-    public Vector3 Diagonal
-    {
-      get { return new Vector3( m_data1[ 0 ], m_data2[ 1 ], m_data2[ 2 ] ); }
-    }
-
-    private Vector2Int GetIndex( int i, int j )
-    {
-      var globalIndex = i + j;
-      if ( globalIndex == 2 && i == 1 )
-        globalIndex = 5;
-      return new Vector2Int( globalIndex % 3, globalIndex / 3 );
-    }
-
-    private Vector3 m_data1;
-    private Vector3 m_data2;
-  }
-
+  /// <summary>
+  /// Optional element "inertial" defining mass properties of a "link".
+  /// This element reads:
+  ///   - Optional element "origin" defining the pose of the inertial reference frame relative
+  ///     to the reference frame of the "link" (<see cref="Pose"/>).
+  ///   - Required element "mass" with required attribute "value". Default: 0.0.
+  ///   - Required element "inertia" with required attributes "ixx", "ixy", "ixz",
+  ///     "iyy", "iyz" and "izz" (<see cref="Inertia"/>). 
+  /// </summary>
   public class Inertial : Pose
   {
+    /// <summary>
+    /// Reads optional "inertial".
+    /// </summary>
+    /// <param name="element">Optional element "inertial" under a "link".</param>
+    /// <returns>Inertial instance if <paramref name="element"/> != null, otherwise null.</returns>
     public static Inertial ReadOptional( XElement element )
     {
       if ( element == null )
@@ -88,9 +26,21 @@ namespace AGXUnity.IO.URDF
       return new Inertial( element );
     }
 
+    /// <summary>
+    /// Mass if the link if "inertial" is defined. Default: 0.0.
+    /// </summary>
     public float Mass { get; private set; } = 0.0f;
+
+    /// <summary>
+    /// Inertia of the link if "inertial" is defined. Default: zeros.
+    /// </summary>
     public Inertia Inertia { get; private set; } = Inertia.zero;
 
+    /// <summary>
+    /// Reads optional element "inertial".
+    /// </summary>
+    /// <param name="element">Optional element "inertial" under "link".</param>
+    /// <param name="optional">Unused.</param>
     public override void Read( XElement element, bool optional = true )
     {
       base.Read( element, true );
@@ -99,6 +49,10 @@ namespace AGXUnity.IO.URDF
       Inertia = Inertia.Read( element?.Element( "inertia" ) );
     }
 
+    /// <summary>
+    /// Construct given optional element "inertial".
+    /// </summary>
+    /// <param name="element">Optional element "inertial" under "link".</param>
     public Inertial( XElement element )
     {
       Read( element );
