+++ AGXUnity/IO/URDF/Utils.cs
@@ -7,6 +7,13 @@ namespace AGXUnity.IO.URDF
 {
   public static class Utils
   {
+    /// <summary>
+    /// Finds line number of the current XElement or XAttribute if
+    /// this feature has been enabled when loading the XDocument.
+    /// </summary>
+    /// <typeparam name="T">XObject type.</typeparam>
+    /// <param name="xmlObject">Current object.</param>
+    /// <returns>A string with "line(X)" where X is the line number.</returns>
     public static string GetLineInfo<T>( T xmlObject )
       where T : XObject
     {
@@ -15,6 +22,16 @@ namespace AGXUnity.IO.URDF
       return $"line({((IXmlLineInfo)xmlObject).LineNumber})";
     }
 
+    /// <summary>
+    /// Finds attribute of given <paramref name="attributeName"/> in <paramref name="element"/>.
+    /// If <paramref name="optional"/> == false and the attribute isn't present in
+    /// <paramref name="element"/> an UrdfIOException is thrown.
+    /// </summary>
+    /// <param name="element">Current element.</param>
+    /// <param name="attributeName">Name of the attribute.</param>
+    /// <param name="optional">False to throw an exception if <paramref name="attributeName"/> isn't
+    ///                        present in <paramref name="element"/>.</param>
+    /// <returns>Attribute in <paramref name="element"/> with given <paramref name="attributeName"/>.</returns>
     public static XAttribute GetAttribute( XElement element,
                                            string attributeName,
                                            bool optional )
@@ -26,6 +43,16 @@ namespace AGXUnity.IO.URDF
       return attribute;
     }
 
+    /// <summary>
+    /// Reads <paramref name="attributeName"/> from <paramref name="element"/> as
+    /// an UnityEngine.Color, expecting four space separated values.
+    /// </summary>
+    /// <param name="element">Current element.</param>
+    /// <param name="attributeName">Color attribute name.</param>
+    /// <param name="defaultColor">Default value if the attribute is optional.</param>
+    /// <param name="optional">False to throw an exception if <paramref name="attributeName"/> isn't
+    ///                        present in <paramref name="element"/>.</param>
+    /// <returns>Read color or <paramref name="defaultColor"/> if the attribute isn't present.</returns>
     public static Color ReadColor( XElement element,
                                    string attributeName,
                                    Color defaultColor,
@@ -44,6 +71,15 @@ namespace AGXUnity.IO.URDF
                         Convert.ToSingle( values[ 3 ] ) );
     }
 
+    /// <summary>
+    /// Reads <paramref name="attributeName"/> from <paramref name="element"/> as
+    /// an UnityEngine.Vector3, expecting three space separated values.
+    /// </summary>
+    /// <param name="element">Current element.</param>
+    /// <param name="attributeName">Attribute name to read from <paramref name="element"/>.</param>
+    /// <param name="optional">False to throw an exception if <paramref name="attributeName"/> isn't
+    ///                        present in <paramref name="element"/>.</param>
+    /// <returns>Read Vector3 or Vector3.zero if the attribute isn't present.</returns>
     public static Vector3 ReadVector3( XElement element,
                                        string attributeName,
                                        bool optional = true )
@@ -60,6 +96,14 @@ namespace AGXUnity.IO.URDF
                           Convert.ToSingle( values[ 2 ] ) );
     }
 
+    /// <summary>
+    /// Reads <paramref name="attributeName"/> from <paramref name="element"/> as a float.
+    /// </summary>
+    /// <param name="element">Current element.</param>
+    /// <param name="attributeName">Attribute name to read from <paramref name="element"/>.</param>
+    /// <param name="optional">False to throw an exception if <paramref name="attributeName"/> isn't
+    ///                        present in <paramref name="element"/>.</param>
+    /// <returns>Read float or 0.0f if the attribute isn't present.</returns>
     public static float ReadFloat( XElement element,
                                    string attributeName,
                                    bool optional = true )
@@ -70,6 +114,14 @@ namespace AGXUnity.IO.URDF
       return (float)attribute;
     }
 
+    /// <summary>
+    /// Reads <paramref name="attributeName"/> from <paramref name="element"/> as a string.
+    /// </summary>
+    /// <param name="element">Current element.</param>
+    /// <param name="attributeName">Attribute name to read from <paramref name="element"/>.</param>
+    /// <param name="optional">False to throw an exception if <paramref name="attributeName"/> isn't
+    ///                        present in <paramref name="element"/>.</param>
+    /// <returns>Read string or string.Empty if the attribute isn't present.</returns>
     public static string ReadString( XElement element,
                                      string attributeName,
                                      bool optional = true )
